requirements.txt
python-crontab==3.0.0
docker
JPype1
confluent-kafka